Add version & run Cobra commands
