+++ cmd/main.go
@@ -1,0 +1,29 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/onflow/flow-evm-gateway/cmd/run"
+	"github.com/onflow/flow-evm-gateway/cmd/version"
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "",
+	Short: "Utility commands for the EVM Gateway Node",
+}
+
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
+}
+
+func main() {
+	rootCmd.AddCommand(version.Cmd)
+	rootCmd.AddCommand(run.Cmd)
+
+	Execute()
+}
